# Conflicts:
#	clientAPI/.idea/workspace.xml

+++ productAPI/src/main/java/com/letscode/store/controller/ProductController.java
@@ -1,13 +1,13 @@
 package com.letscode.store.controller;
 
 import com.letscode.store.dto.ProductDTO;
-import com.letscode.store.model.Product;
+import com.letscode.store.exception.AlreadyExistException;
+import com.letscode.store.exception.NotAuthorizedException;
 import com.letscode.store.service.ProductService;
-import com.querydsl.core.types.Predicate;
+import com.letscode.store.service.TokenService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.querydsl.binding.QuerydslPredicate;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 
@@ -20,26 +20,39 @@ public class ProductController {
 
     private final ProductService productService;
 
+    private final TokenService tokenService;
+
     @ResponseStatus(code = HttpStatus.CREATED)
     @PostMapping
-    public ProductDTO saveProduct(@RequestBody @Valid ProductDTO productDTO){
-
+    public ProductDTO saveProduct(@RequestBody @Valid ProductDTO productDTO, @RequestHeader("authorization") String token) throws AlreadyExistException {
+        if (tokenService.getAuthenticate(token)){
         return productService.saveProduct(productDTO);
+        }
+        throw new NotAuthorizedException("não permitido");
     }
 
     @GetMapping
-    public Page<ProductDTO> listProduct(@QuerydslPredicate(root = Product.class) Predicate predicate, Pageable pageable){
-        return productService.listProduct(predicate, pageable);
+    public Page<ProductDTO> listProduct(ProductDTO productDTO, Pageable pageable, @RequestHeader("authorization") String token) {
+        if (tokenService.getAuthenticate(token)){
+        return productService.listProduct(productDTO, pageable);
+        }
+        throw new NotAuthorizedException("não permitido");
     }
 
     @PatchMapping("/{code}")
-    public ProductDTO updateProduct(@RequestBody @Valid ProductDTO productDTO) {
-        return productService.updateProduct(productDTO);
+    public ProductDTO updateProduct(@RequestBody @Valid ProductDTO productDTO, @RequestHeader("authorization") String token) {
+        if (tokenService.getAuthenticate(token)){
+        return productService.updateProduct(productDTO);      }
+        throw new NotAuthorizedException("não permitido");
     }
 
 
     @DeleteMapping("/{code}")
-    public void deleteClient(@PathVariable String code) {
-        productService.deleteProduct(code);
+    public void deleteProduct(@PathVariable String code, @RequestHeader("authorization") String token) {
+        if (tokenService.getAuthenticate(token)){
+            productService.deleteProduct(code);
+        }else{
+            throw new NotAuthorizedException("não permitido");
+        }
     }
 }
